+++ CHANGELOG.md
@@ -12,6 +12,11 @@ PDBTools.jl Changelog
 
 Version 3.33.3-DEV
 --------------
+- ![FEATURE][badge-feature] Add `set_phi!` and `set_psi!` to set backbone dihedral angles of a protein structure.
+- ![FEATURE][badge-feature] Add `extended_chain` and `extended_chain!` to build the fully-extended (all-trans) conformation of a protein.
+- ![FEATURE][badge-feature] Add `*`, `+`, and `-` to scale and combine `TransferFreeEnergy` objects, used to simplify the `mvalue(::MTRecordDenaturedModel, ...)` implementation.
+- ![EXPERIMENTAL][badge-experimental] Implement `MTRecord` transfer model to compute m-values and transfer free energies, in urea or betaine.
+- ![EXPERIMENTAL][badge-experimental] Document the `MTRecord` model.
 
 Version 3.33.2
 --------------
@@ -113,7 +118,7 @@ Version 3.23.1
 
 Version 3.23.0
 --------------
-- ![FEATURE][badge-feature] Add `CreamerDenaturedModel` type, and `mvalue(::CreamerDenaturedModel)` to comnpute m-values associated with denaturation directly.
+- ![FEATURE][badge-feature] Add `CreamerDenaturedModel` type, and `mvalue(::CreamerDenaturedModel)` to compute m-values associated with denaturation directly.
 
 Version 3.22.0
 --------------
@@ -143,7 +148,7 @@ Version 3.20.0
 --------------
 - ![FEATURE][badge-feature] Add support for `filter(f, ::HBonds)` and slicing `HBonds` objects with `hbonds[1:5]`, or with index vectors.
 - ![FEATURE][badge-feature] Add iteration interface (`iterate`, `keys`, `eachindex`) for `HBonds` data structure, enabling `for` loops, `collect`, `findfirst`, etc.
-- ![INFO][badge-info] Document comprehensivelly the hydrogen-bonds functionality.
+- ![INFO][badge-info] Document comprehensively the hydrogen-bonds functionality.
 - ![INFO][badge-info] Code formatting and small adjustments.
 - ![INFO][badge-info] Replace BenchmarkTools.jl by Chairmarks.jl in allocation tests.